generate __data.js
